document-api/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Pydantic for configuration (compatible with Python 3.13)
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Google Cloud libraries
google-cloud-storage>=2.14.0
google-cloud-discoveryengine>=0.11.11
google-api-core>=2.15.0
google-auth>=2.27.0

# Database
asyncpg>=0.29.0

# Additional utilities
python-dotenv>=1.0.0
